move pools fuzzing to hongfuzz
